Only use 'd' suffix when debug libraries are present.
Recent versions of Debian and Ubuntu (beginning with Stretch and Bionic
respectively) do not provide separate debug library versions of Poco.

The refactored debug check now actually verifies that a d-suffixed
library exists. If not it falls back to using the non-suffixed version
of the library which may or may not include debug symbols.